requirements.txt
rembg[new]
Pillow
